Añado la logica para el riego automatico y la iluminacion automatica

+++ esp32_sotfware/include/sistem_control.cpp
@@ -21,24 +21,69 @@ void receptionSystem( String topic, String message){
 }
 
 
-void taskWater( int pin){};
-void taskLight( int pinStart,  int pinEnd){};
-void taskSystem( JsonObjectConst User ){
+void taskWater( int pin_humCap, int pin_humEc,int limit, int pinout){  Serial.println("Activado el sistema de riego"); };
+void taskLight( int led_start,  int led_end, int r, int g, int b){ Serial.println("Activado el sistema de iluminacion"); };
+void taskSystem( JsonObjectConst plant ){
  
  struct tm timeinfo;
   if(!getLocalTime(&timeinfo)){
     Serial.println("Failed to obtain time");
     return;
   }
- taskWater( User["Water"]["pin"] );
+    
+  //Observamos el entorno por cada planta
+  // Last_time_water : "YYYY:MM:DDD"
+  time_t timeinfo_2 = mktime(&timeinfo);
 
- if( ((timeinfo.tm_hour >= User["Light"]["startHour"]) && (timeinfo.tm_min >= User["Light"]["startMin"])) &&
-     ((timeinfo.tm_hour >= User["Light"]["endHour"]) && (timeinfo.tm_min >= User["Light"]["endMin"])) 
-  ){
+  double diff = difftime (timeinfo_2, (plant["water"]["last_water"].as<unsigned long>()) );
 
-    taskLight( User["Light"]["pinStart"] , User["Light"]["pinEnd"]);
+
+  if( (diff/(60*60*60*24))  >=  plant["water"]["frequency"].as<String>().toInt() ) {
+    
+    taskWater( plant["pinout"]["humCap"].as<int>(),
+               plant["pinout"]["humEC"].as<int>(),
+               plant["water"]["limit"].as<int>(),
+               plant["water"]["pinout"].as<int>());
+  
   }
 
+
+  if(    (timeinfo.tm_hour <= (plant["light"]["time_start"].as<String>()).substring (0 , 2).toInt()) 
+      && (timeinfo.tm_min  <= (plant["light"]["time_start"].as<String>()).substring (3 , 5).toInt()) 
+      && (timeinfo.tm_hour >= (plant["light"]["time_stop"].as<String>()).substring (0 , 2).toInt()) 
+      && (timeinfo.tm_min  >= (plant["light"]["time_stop"].as<String>()).substring (3 , 5).toInt()) 
+  ){
+
+    taskLight( plant["light"]["led_start"].as<int>(), 
+               plant["light"]["led_end"].as<int>(),
+               plant["light"]["color_red"].as<int>(),
+               plant["light"]["color_green"].as<int>(),
+               plant["light"]["color_blue"].as<int>());
+  }
+  
+  if( debugging ){
+    Serial.println("------------------  taskSystem   ---------------------------------");
+    
+    Serial.print("Time Hours: "); Serial.println( timeinfo.tm_hour );
+    Serial.print("Time min: "); Serial.println( timeinfo.tm_min );
+    Serial.print("ID plant: "); Serial.println( plant["info"]["date"].as<int>());
+
+    Serial.print("[light][time_start]: "); Serial.println( plant["light"]["time_start"].as<String>() );
+    Serial.print("[light][time_start][Hour]: "); Serial.println( (plant["light"]["time_start"].as<String>()).substring (0 , 2).toInt()  );
+    Serial.print("[light][time_start][min]: "); Serial.println(  (plant["light"]["time_start"].as<String>()).substring (3 , 5).toInt() );
+   
+    Serial.print("[light][time_stop]: "); Serial.println( plant["light"]["time_stop"].as<String>()  );
+    Serial.print("[light][time_stop][Hour]: "); Serial.println( (plant["light"]["time_stop"].as<String>()).substring (0 , 2).toInt()  );
+    Serial.print("[light][time_stop][min]: "); Serial.println(  (plant["light"]["time_stop"].as<String>()).substring (3 , 5).toInt() );
+
+    Serial.print("[water][last_water]: "); Serial.println( plant["water"]["last_water"].as<unsigned long>() );
+    Serial.print("diff: "); Serial.println( diff/(60*60*60*24) );
+    Serial.print("[water][frequency]: "); Serial.println( plant["water"]["frequency"].as<String>() );
+    
+    Serial.println("------------------  taskSystem end   ---------------------------------");
+  }
+    
+  
 
 }
 void tasksControl(){
@@ -114,40 +159,59 @@ bool getUser(){
   if( wifi_status() ){
     HTTPClient http;
 
-    http.begin( (urlGetUser+getMac()) );
-    int httpCode = http.GET();  // Realizar petición
-  
-    if (httpCode > 0) {
+    Serial.print("search user: "); Serial.println(urlGetUser+getMac() );
+    if( http.begin(espClient, urlGetUser+getMac() )){
+      Serial.print("[HTTP] GET...\n");
+      int httpCode = http.GET();  // Realizar petición
 
-      if (httpCode == HTTP_CODE_OK || httpCode == HTTP_CODE_MOVED_PERMANENTLY) {
+      if (httpCode > 0) {
+        Serial.printf("[HTTP] GET... code: %d\n", httpCode);
         
-        String payload = http.getString();   // Obtener respuesta
-        auto error = deserializeJson(localUser, payload);
+        if (httpCode == HTTP_CODE_OK || httpCode == HTTP_CODE_MOVED_PERMANENTLY) {
         
-        if( error == DeserializationError::Ok ){  
+          String payload = http.getString();   // Obtener respuesta
+          auto error = deserializeJson(localUser, payload);
+        
+          if( error == DeserializationError::Ok ){  
           
-          if( payload != readFile(SD, SD_path_user)){
-            Serial.println("Different user both cloud and local");
-            deleteFile(SD, SD_path_user);
-            writeFile(SD, SD_path_user, payload.c_str() );
+            // If we can't fount the user in local. Go to sleep
+            if(!SD.begin()){
+              Serial.println("Card Mount Failed");
+              sisError(5);
+            }else{
+              listDir(SD, "/", 0);
+              if ( SD.exists(SD_path_user) ){
+                if( payload != readFile(SD, SD_path_user)){
+                  Serial.println("Different user both cloud and local");
+                  deleteFile(SD, SD_path_user);
+                  delay(100);
+                  writeFile(SD, SD_path_user, payload.c_str() );
+                }
+              }else{
+                writeFile(SD, SD_path_measure, payload.c_str() );
+              }
+            }
+            SD.end();
+            //if(debugging){ Serial.println(payload); };   // Mostrar respuesta por serial
+            UserFinded = true;
           }
-          
-          if(debugging){ Serial.println(payload); };   // Mostrar respuesta por serial
-          UserFinded = true;
+        }else if(httpCode == 204){
+          sisError (3); // 3: User don't exist
+          Serial.print("search user: "); Serial.println(getMac());
+          Serial.println("User don't exist in the server");
+        }else{
+          sisError (4); // 4: Bad comunication in http
         }
-      }else if(httpCode == 204){
-        sisError (3); // 3: User don't exist
-        Serial.println("User don't exist");
-      }else{
-        sisError (4); // 4: Bad comunication in http
+      }else {
+        sisError (2, http.errorToString(httpCode).c_str() ); // 2: impossible to reach the server
       }
-    }else {
-      sisError (2, http.errorToString(httpCode).c_str() ); // 2: impossible to reach the server
-    }
   
-    http.end();
-  }
+      http.end();
 
+    }else {
+      Serial.printf("[HTTP} Unable to connect\n");
+    }
+  }
   return UserFinded;  
 }
